+++ lib/src/model/gen_image/request/edit_file.dart
@@ -59,14 +59,14 @@ class EditImageRequest {
 
   Future<FormData> convert() async {
     return FormData.fromMap({
-      'image': File(image.path).existsSync()
+      'image': (await File(image.path).exists())
           ? await MultipartFile.fromFile(
               image.path,
               filename: image.name,
               contentType: MediaType('image', 'png'),
             )
           : null,
-      'mask': !File('${mask?.path}').existsSync()
+      'mask': !(await File('${mask?.path}').exists())
           ? null
           : await MultipartFile.fromFile(
               '${mask?.path}',